# api/docs/swagger.yaml
...
host: api.utdnebula.com
info:
  contact: {}
  description: The developer Nebula Labs API for access to pertinent UT Dallas data
  title: dev-nebula-api
  version: 1.0.0
paths:
...
  /swagger/{file}:
    get:
      description: Returns the OpenAPI/swagger spec for the API
      operationId: swagger
      parameters:
      - description: The swagger file to retrieve
        in: path
        name: file
        required: true
        type: string
      responses:
        "200":
          description: OK
      security: []
schemes:
- http
- https
security:
- api_key: []
securityDefinitions:
  api_key:
    in: header
    name: x-api-key
    type: apiKey
  storage_key:
    in: header
    name: x-storage-key
    type: apiKey
swagger: "2.0"
...